保存进度

+++ LumiRender/src/core/refl/factory.h
@@ -90,14 +90,14 @@ namespace luminous {
 #define SET_PTR_VALUE(ptr, offset, val) (reinterpret_cast<luminous::ptr_t*>(&((reinterpret_cast<std::byte*>(ptr))[offset])))[0] = ptr_t(val)
 
         template<typename T>
-        LM_NODISCARD luminous::ptr_t get_ptr_value(T ptr, uint32_t offset) {
+        LM_NODISCARD luminous::ptr_t get_ptr_value(T ptr, uint64_t offset) {
             static constexpr bool condition = std::is_pointer_v<T> || std::is_same_v<luminous::ptr_t, T>;
             static_assert(condition, "the type T must be ptr_t or pointer!");
             return GET_PTR_VALUE(ptr, offset);
         }
 
         template<typename T, typename U>
-        INLINE void set_ptr_value(T ptr, uint32_t offset, U val) {
+        INLINE void set_ptr_value(T ptr, uint64_t offset, U val) {
             static constexpr bool condition1 = std::is_pointer_v<T> || std::is_same_v<luminous::ptr_t, T>;
             static constexpr bool condition2 = std::is_pointer_v<U> || std::is_same_v<luminous::ptr_t, U>;
             static_assert(condition2 && condition1, "the type T and U must be ptr_t or pointer!");
